Moved Comment component to a high lever order
