+++ React/wbooks/src/app/components/Comment/styles.scss
@@ -1,4 +1,4 @@
-@import '../../../../../constants/colors.scss';
+@import '../../../constants/colors.scss';
 
 .commentContainer {
   display: flex;
